提交减缩版的GPUimage

+++ framework/Source/iOS/GPUImagePicture.m
@@ -144,7 +144,7 @@
                 CGBitmapInfo byteOrderInfo = bitmapInfo & kCGBitmapByteOrderMask;
                 if (byteOrderInfo == kCGBitmapByteOrder32Little) {
                     /* Little endian, for alpha-first we can use this bitmap directly in GL */
-                    CGImageAlphaInfo alphaInfo = bitmapInfo & kCGBitmapAlphaInfoMask;
+                    CGImageAlphaInfo alphaInfo = (uint32_t)bitmapInfo & (uint32_t)kCGBitmapAlphaInfoMask;
                     if (alphaInfo != kCGImageAlphaPremultipliedFirst && alphaInfo != kCGImageAlphaFirst &&
                         alphaInfo != kCGImageAlphaNoneSkipFirst) {
                         shouldRedrawUsingCoreGraphics = YES;
@@ -152,7 +152,7 @@
                 } else if (byteOrderInfo == kCGBitmapByteOrderDefault || byteOrderInfo == kCGBitmapByteOrder32Big) {
 					isLitteEndian = NO;
                     /* Big endian, for alpha-last we can use this bitmap directly in GL */
-                    CGImageAlphaInfo alphaInfo = bitmapInfo & kCGBitmapAlphaInfoMask;
+                    CGImageAlphaInfo alphaInfo = (uint32_t)bitmapInfo & (uint32_t)kCGBitmapAlphaInfoMask;
                     if (alphaInfo != kCGImageAlphaPremultipliedLast && alphaInfo != kCGImageAlphaLast &&
                         alphaInfo != kCGImageAlphaNoneSkipLast) {
                         shouldRedrawUsingCoreGraphics = YES;
